Global: Ensure aspect ratios are floats everywhere
- This was something I had noticed yesterday while implementing
  `forceAspectRatio` but didn't want to spend the tedious time searching
  for all the code and fixing it everywhere. But it's a BUG, so it was
  still annoying me today that I had left it in there... I've now fixed
  it properly by searching for all aspect ratio code via `ag
  'width.*\/.*height'`, which seems to have found them all. The problem
  with the old code was that `int(1080) / int(1080) = int(1)`, in other
  words every time that you divide equal width and height integers in
  PHP you always get an int instead of a float. That meant that
  functions which claimed to return a float aspect ratio were lying in
  case of square images. And MOST SERIOUSLY it also meant that strict
  `=== 1.0` comparisons were impossible, since the variable became an
  unexpected int every time the aspect ratio was perfectly square.

- Fixed it by always casting the result to a float *everywhere*, so that
  `(float) (int(1080) / int(1080)) = float(1.0)`. Even added this strict
  casting in places where we don't need to do casting (such as in printf
  statements), just so that nobody sees any bugged non-casting code
  anywhere and accidentally copies the bug into a serious code location
  again in the future. Because it is a very evil and well-hidden bug, so
  hopefully the copious amounts of `(float)` conversion and new docs
  will now alert future readers that something special is going on with
  that code...

- Also fixed some incorrect PHPdoc `@return` values. And merged two
  redundant lines in `examples/uploadAlbum.php`.

+++ src/Media/Geometry/Dimensions.php
@@ -25,7 +25,8 @@ class Dimensions
     {
         $this->_width = (int) $width;
         $this->_height = (int) $height;
-        $this->_aspectRatio = $this->_width / $this->_height;
+        // NOTE: MUST `float`-cast to FORCE float even when dividing EQUAL ints.
+        $this->_aspectRatio = (float) ($this->_width / $this->_height);
     }
 
     /**
@@ -51,7 +52,7 @@ class Dimensions
     /**
      * Get stored aspect ratio for these dimensions.
      *
-     * @return int
+     * @return float
      */
     public function getAspectRatio()
     {
